requirements.txt
aiogram==2.25.2
uvicorn
fastapi
git+https://github.com/Qewertyy/LexicaAPI/
python-dotenv